Se agregan mas test y bd

+++ Punto_Venta/src/main/java/mx/aragon/unam/service/producto/categorias/CategoriaServiceImpl.java
@@ -27,12 +27,12 @@ public class CategoriaServiceImpl implements CategoriaService {
     }
 
     @Override
-    public void deleteById(Short id) {
+    public void deleteById(Integer id) {
         categoriaRepository.deleteById(id);
     }
 
     @Override
-    public CategoriaEntity findById(Short id) {
+    public CategoriaEntity findById(Integer id) {
         return categoriaRepository.findById(id).orElse(null);
     }
 }
